Display the welcome message via getty's issue file
The welcome message with the FOSS-Cloud ASCII logo will now be displayed
by the getty processes on all terminals instead of only once after
booting. The welcom.sh script now writes the /etc/issue file which in
turn will be read, interpreted and displayed by agetty. Because of
agetty issue escapes, a logo which has all backslashes (\) escaped was
necessary. This is the first part of the fix for bug #69.